pyproject.toml
[project]
name = "adk-hybrid-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.15.1",
]
